+++ pyspelling/filters/__init__.py
@@ -238,7 +238,7 @@ class Filter(plugin.Plugin):
     def filter(self, source_file, encoding):  # noqa A001
         """Open and filter the file from disk."""
 
-        with codecs.open(source_file, 'r', encoding=encoding) as f:
+        with open(source_file, 'r', encoding=encoding, errors='strict') as f:
             text = f.read()
         return [SourceText(text, source_file, encoding, 'text')]