+++ pyspelling/filters/cpp.py
@@ -445,7 +445,7 @@ class CppFilter(filters.Filter):
     def filter(self, source_file, encoding):  # noqa A001
         """Parse CPP file."""
 
-        with codecs.open(source_file, 'r', encoding=encoding) as f:
+        with open(source_file, 'r', encoding=encoding, errors='strict') as f:
             text = f.read()
 
         return self._filter(text, source_file, encoding)