+++ pyspelling/filters/markdown.py
@@ -1,6 +1,5 @@
 """Markdown filter."""
 from .. import filters
-import codecs
 import markdown
 
 
@@ -37,7 +36,7 @@ class MarkdownFilter(filters.Filter):
     def filter(self, source_file, encoding):  # noqa A001
         """Parse Markdown file."""
 
-        with codecs.open(source_file, 'r', encoding=encoding) as f:
+        with open(source_file, 'r', encoding=encoding, errors='strict') as f:
             text = f.read()
         return [filters.SourceText(self._filter(text), source_file, encoding, 'markdown')]