+++ pyspelling/filters/url.py
@@ -1,6 +1,5 @@
 """URL filter."""
 from .. import filters
-import codecs
 import re
 
 # Bare link/email detection
@@ -63,7 +62,7 @@ class URLFilter(filters.Filter):
     def filter(self, source_file, encoding):  # noqa A001
         """Open and filter the file from disk."""
 
-        with codecs.open(source_file, 'r', encoding=encoding) as f:
+        with open(source_file, 'r', encoding=encoding, errors='strict') as f:
             text = f.read()
 
         return [filters.SourceText(self._filter(text), source_file, encoding, 'url-free')]